Drop libcurl support (#10)
summary: Recent releases of libcurl broke our static build of htslib. I installed an older release of libcurl, but this didn't fix the problem. Instead of continuing to troubleshoot, we decided to disable libcurl support for this Windows version of htslib (thus disabling S3 support for bcftools and any other tools that rely on htslib for this feature).